lime-proto-bmx6: simplify bmx6 config, merge local[4|6] tunIns into tun[4|6]Address
bmx6 automatically announces the correspondent tunIn
if tunXAddress is a network (/64) rather than a host (/128).
So it's better to not announce specific host routes,
since this creates many extra routes and tunnels
which clutter up the system.

This commit is a combined revert of both:
d1b2e45 lime-proto-bmx6: fix bmx6 main addresses
9bb58e2 lime-proto-bmx6 Add also local ipv6 announcement