Introduce new package smonit
Small and modular monitor for daemons.
It executes every 5 minutes and run a set of hooks to check the
daemons and status of the system. Hooks are installed on /etc/smonit and
must define a function for quickfix (like daemon restart) and longfix
(like rebuild configuration for daemon). Longfix is launched only if
after 3 quickfixes the problem is not solved. Logs are stored on /tmp/smonit/<daemon_name>.

First commit with dnsmasq and bmx6 hooks.

Signed-off-by: Pau Escrich <p4u@dabax.net>