Add anygw_mac_enable option on lime config
Adds the possibility to disable anyMAC from anyGW, so the nodes are not
sharing the same MAC address anymore. This should fix unexpected
behaviors on some routers (those with internal switch). By default
anyMAC will be disabled, it might be enabled for a more agressive
roaming of the network, which might be useful on specific deployments.

Signed-off-by: Pau Escrich <p4u@dabax.net>