firewall: add lm_* interfaces to 'lan' zone
To get the stock firewall working nicely, add a all lm_* interfaces
to the 'lan' zone. And we really shouldn't change the defaults for
the wan zone (!), why ever this was done, it really shouldn't.
If the wan interface isn't used as a wan interface, it shouldn't be
called wan -- calling it wan and opening the firewall for all
incoming packages while keeping SSH with no-password-auth open is
real madness and will get people pwned.