Fix travis fail on fork pull request
Pull requests sent from forked repositories do not have access to encrypted variables or data.
As you can not access the variables you can not decrypt the key and perform the tasks in after_success.sh
This change verifies that the action comes from the libremesh/lime-packages repository and not from a fork.