lime-cc: add package
lime-cc has two parts one is lime-cc-client and one is lime-cc-server

* lime-cc-client
check every 5 minutes for a new configuration received by bmx7-sms.
if the configuration is valid and the node is trusted the configuration
is used as default, lime-config and lime-apply runs.
if not, the configuration is ignored.

* lime-cc-server
distributes it own lime-defaults to the cloud. If changes happen to the
config it will be distributed to the cloud.

this is a prove of concept for one of the approaches mentioned
[here](https://github.com/libremesh/lime-packages/issues/304)

bmx7-sms is required to make this working

lime-cc: use uci options

uci option to set trusted nodes in lime-defaults

config trusted_nodes 'trusted_nodes'
    list node_id '<node_id1>'
    list node_id '<node_id2>'
    list node_id '<node_id3>'

if a lime-defaults file is received it will automatically overwrite the
existsting lime-defaults and lime-{config,apply} runs.

lime-cc: rename to lime-bmx7-rc