lime/wireless: activate wifi modes based on suffix
Add the option to add `_2ghz` and `_5ghz` as a suffix to the
activated wifi modes (ap, apname, ieee80211s).

This is usefull as the specific radio configuration (radio0, radio1,
etc) only works with /etc/config/lime but not with lime-defaults.

When using encrypted mesh and encrypted AP it comes to an error
mentioned [here](https://github.com/libremesh/lime-packages/issues/208).