lime-system: add function arrayConcat to fix wifi
Currently setting specific radio options lead to a total overwrite of
all settings provided in lime and lime-defaults.

Changing just the channel results in a bad config as ssid, etc. are empty.

Instead, this patch introduces the arrayConcat function where the
specific settings are attached to the defaults and so overwrite them.

In client mode no defaults are used but only specific channels. If not
given the fallback option is "auto".

Signed-off-by: Paul Spooren <spooren@informatik.uni-leipzig.de>