utils: introduce %m in IPs to use macid
macid is like netid a md5sum hash of mac. The idea is to have static IP
addresses for devices which sustain after a reset or reflash, while
keeping collision chance as lower as using %M aka the real mac.
This is due to the fact that when using similar devices, they tend to
have nearly equal mac addresses.

A possible network could look like this: 10.%m1.%m2.%m3/25 which would
result in a collision chance of `n/(2 * 256^2)` where `n` is the number
of used devices.

Signed-off-by: Paul Spooren <spooren@informatik.uni-leipzig.de>