utils: add function tableMelt
I don't know what lead to the arrayContact function but it can't be used
in that case: `options` are loaded via `config.get_all()` and result in
an dict like table. Specific options are also a dict and can't be added
as an array.

New function tableMelt returns the combination of both received tables,
where values of the second table are used when exist in both tables.

Also both function return the value rather then manipulating the
incomming table. Therebey wireless.lua is adjusted.

Signed-off-by: Paul Spooren <mail@aparcar.org>