lime-system: use lime-profile if available
Network profiles are somewhat popular, but cause problems when they're
used as packages, as lime-system and the packaged profile both supply
the file `/etc/config/lime-defaults`. A solution is to introduce
`lime-profile` which is then used instead of lime-defaults.

This has the advantage that new releases with different defaults don't
break compatiblity.

Signed-off-by: Paul Spooren <mail@aparcar.org>