Introduce new package bmx7-mdns
bmx7 mesh DNS is a tool to provide simple distributed DNS functionality to the mesh network.
Uses bmx7-sms plugin to announce domains and IPs and dnsmasq for local resolve.
TLD are by default ".mesh" and ".mesh6" but can be changed, not matching TLDs are ignored (to avoid poisoning).
Node hostname is automatically announced, but also other domains can be published using syntax name.tld@IP
Comes with a simple LUCI interface to monitor announced domains and add new ones to the mesh network.
It's based on 2013 qmp.cat mDNS for bmx6 code.

Signed-off-by: p4u <p4u@dabax.net>