testing: add dockerfile for unittesting
Based on a (very small!) Alpine Linux image with lua 5.1 and
luarocks installed.
As some openwrt modules are needed or extremely helpful, eg uci,
libjson-c, luci-lib-ip, nixio, they are installed directly from
the 18.06.4 x86_64 packages. The rationale is to have a minimal
maintenance but at the same time do not reinvent the wheel.
Some modules should be faked instead of installed, eg: iwinfo.