lime.wireless: add dev.per_band_index attribute to scandevices() result
The idea is to have a way to enumerate the devices per band. This uses
the number in the name of the device for the sorting order. So radio0
will always have a per_band_index = 0, and radio1 will be 1 if it is
of the same band as radio0 or 0 if it is of a different band.