add deferable-reboot package
This package provides a deferable reboot mechanism. The system
will be rebooted when the uptime of the system reach a specified amount
of seconds and if /tmp/deferable-reboot.defer file does not contain a
number that is greater than the current uptime.
This way if anyone wants to postpone the reboot, the way is to store the
next checkpoint in seconds. For example if the system must not reboot from
now and up to 120 seconds in the future then something like this can be donde
awk '{print $1 + 120}' /proc/uptime > /tmp/deferable-reboot.defer