anygw: use dnsmasq host-record instead of address
This fixes dnsmasq returning the anycast address when the dns query
is for a host if dnsmasq does not know about it. For example
pepe.thisnode.info was returning the anycast address if pepe was not
known by dnsmasq (or only the ipv4 was known and the ipv6 was
requested).

This is related to openwrt's:
dfea3bae11dd5c207182371ce1fdca763fb5bbe0
    Author: Jo-Philipp Wich <jow@openwrt.org>
    Date:   Mon Jun 17 11:55:30 2013 +0000

        dnsmasq: use host-record instead of address

        Using "--address" for individual host A records is broken, use
	"--host-record" instead.