CI: build all packages via GitHub action
Build all packages of the lime-feed when commits are pushed to master
branch or a new tag/release is added starting with `v`.

Packages are then uploaded to `libremesh/lime-feed` usable as GitHub
page. This can be extended to use a remote S3 storage or even rsync if
desired.

Via the secret KEY_BUILD all packages are signed.

Signed-off-by: Paul Spooren <mail@aparcar.org>