tinc: remove package
The package should be offered upstream to support multiple
architectures, as this is not part of the LiMe build system.

It was initially introduced for LibreNet6, which required tinc 1.1pre
while packages.git only offers tinc 1.0.36. Instead of having a arch
dependent package in a *archless* feed, remove it now and hope it's
available in packages.git soon.

Signed-off-by: Paul Spooren <mail@aparcar.org>