fbw: use a more permisive temporary wifi config
The temporary wifi config is used to download the config from the
detected remote nodes. Using a restrictive wireless config (high
mcast_rate, short distance) makes this download phase to take more
time or even to fail. Also it does not make a lot of sense for a
node that was already selected as a candidate to download the config.
If we want to restrict the nodes per their signal we should do that
in the scan phase and not in the temporary config.