pirania: set uhttpd http keep alive to 0
This makes redirect script to answer with
`Connection: close` header which in turn
forces next attempt to that url to be another
tcp connection and be tagged by iptables rules
again.

In practice this means that now when you click
on continue, you get redirected to the original url
and not to the portal again.