Update README.adoc
This project is GREAT! I found it from a brief discussion at https://forum.openwrt.org/t/docs-need-review-wiki-article-on-wi-fi-extender-with-relayd/201803/23

I have taken the liberty of making an editorial pass over the README file to fix misspellings ("it's" is always a contraction for "it is" - all other uses are just "its") and clarify some of the language.

Please forgive me for changing the meaning of anything - I didn't mean to. I hope this becomes an important part of OpenWrt.

Best regards,
Rich Brown
Lyme, NH USA