packages: add the package lime-owut
This package, which depends on the package 'owut' from openwrt,
installs an uci-default script to configure owut to also use the
LibreMesh package repositories.

The uci-default script would be in `/rom/etc/uci-defaults/52_lime-owut`
The order of execution of owut-related uci-defaults is this:
```
50-attendedsysupgrade		# by attendedsysupgrade-common
51-attendedsysupgrade-owut	# by owut
52_lime-owut			# by lime-owut
92_add-lime-repos		# by lime-system
```
Where the last `92_add-lime-repos` is the one that adds the apk
repositories and repository_keys for libremesh.